Fixing weird INT_MIN case

+++ holberton.h
@@ -11,7 +11,7 @@ int _putchar(char c);
 int _putc(va_list c);
 int _puts(va_list str);
 int _putmod(void);
-int _pow(int x, int y);
+long int _pow(long int x, long int y);
 int _putint(va_list n);
 int _printf(const char *format, ...);
